+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 Bug Handling
 ************
 
+- Fixed possible bug with using a threading.Condition in the set partitioner.
+  Set partitioner uses new rlock_object handler API to get an appropriate RLock
+  for gevent.
 - Issue #17 fixed: Wrap timeout exceptions with staticmethod so they can be
   used directly as intended. Patch by Bob Van Zant.
 - Fixed bug with client reconnection looping indefinitely using an expired